backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
scikit-learn==1.4.2
shap==0.45.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.0
google-generativeai==0.5.4
scipy==1.13.0
groq==0.9.0
python-dotenv==1.0.1